Forest: extracted this abstraction to its own module
