+++ src/core/Makefile
@@ -1,5 +1,5 @@
 UTILS=transformations basic_utilities task induction_variables loops architecture clean_metadata callgraph scheduler metadata_manager loop_transformer
-ANALYSIS=pdg talkdown alloc_aa dataflow loop_structure invariants
+ANALYSIS=pdg talkdown alloc_aa dataflow loop_structure forest invariants
 ENABLERS=loop_distribution loop_unroll loop_whilifier outliner
 ALL=$(UTILS) $(ANALYSIS) $(ENABLERS) hotprofiler unique_ir_marker noelle scripts
 
@@ -39,6 +39,9 @@ transformations:
 	cd $@ ; ../../scripts/run_me.sh
 
 loop_structure:
+	cd $@ ; ../../scripts/run_me.sh
+
+forest:
 	cd $@ ; ../../scripts/run_me.sh
 
 task:
