+++ src/core/forest/CMakeLists.txt
@@ -1,6 +1,6 @@
 # Project
 cmake_minimum_required(VERSION 3.13)
-project(LoopStructure)
+project(Forest)
 
 # Dependences
 include(${CMAKE_CURRENT_SOURCE_DIR}/../../scripts/DependencesCMake.txt)
@@ -11,9 +11,7 @@ add_subdirectory(src)
 # Install
 install(
   FILES
-  include/noelle/core/LoopStructure.hpp 
-  include/noelle/core/LoopEnvironment.hpp
-  include/noelle/core/EnvBuilder.hpp
+  include/noelle/core/StayConnectedNestedLoopForest.hpp
   DESTINATION 
   include/noelle/core
   )
