+++ src/core/induction_variables/src/CMakeLists.txt
@@ -35,6 +35,7 @@ include_directories(${LLVM_INCLUDE_DIRS}
   ../../pdg/include 
   ../../hotprofiler/include 
   ../../loop_structure/include 
+  ../../forest/include
   ../../talkdown/include
   ../../invariants/include
   ../include
